fix circular dependency warnings by moving enums into their own file
