+++ libs/error-analysis/src/lib/ErrorAnalysisDashboard/ErrorAnalysisEnums.ts
@@ -1,0 +1,26 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export enum ErrorAnalysisOptions {
+  TreeMap = "TreeMap",
+  HeatMap = "HeatMap"
+}
+
+export enum GlobalTabKeys {
+  DataExplorerTab = "DataExplorerTab",
+  GlobalExplanationTab = "GlobalExplanationTab",
+  LocalExplanationTab = "LocalExplanationTab"
+}
+
+export enum ViewTypeKeys {
+  ErrorAnalysisView = "ErrorAnalysisView",
+  ExplanationView = "ExplanationView"
+}
+
+export enum PredictionTabKeys {
+  CorrectPredictionTab = "CorrectPredictionTab",
+  IncorrectPredictionTab = "IncorrectPredictionTab",
+  WhatIfDatapointsTab = "WhatIfDatapointsTab",
+  AllSelectedTab = "AllSelectedTab",
+  InspectionTab = "InspectionTab"
+}
